refactor(services/repositories/models): global refactor

+++ src/main/java/com/chatop/chaTop/repository/RentalRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RentalRepository extends CrudRepository<Rental, Long> {
+public interface RentalRepository extends CrudRepository<Rental, Integer> {
 }
